requirements.txt
anyio==3.7.1
asgiref==3.7.2
bootstrap4==0.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
Django==3.2.20
django-crispy-forms==2.0
django-environ==0.10.0
django-telegrambot==1.0.1
djangorestframework==3.3.2
exceptiongroup==1.1.2
future==0.18.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
psycopg2==2.9.7
python-telegram-bot==6.0.1
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.6
